added a docstring + moved functions to ccc_lmdb
